Rename CompatiblePeripheral to PeripheralType and move scan filter build function
