+++ app/src/main/java/nl/quantifiedstudent/watch/protocol/AbstractBluetoothProtocol.kt
@@ -3,10 +3,13 @@ package nl.quantifiedstudent.watch.protocol
 import android.annotation.SuppressLint
 import android.bluetooth.BluetoothDevice
 import android.bluetooth.BluetoothGatt
+import android.bluetooth.le.ScanFilter
 
 abstract class AbstractBluetoothProtocol : DefaultBluetoothGattCallback() {
 
     protected lateinit var gatt: BluetoothGatt
+
+    abstract val compatiblePeripherals: Collection<PeripheralType>
 
     @SuppressLint("MissingPermission")
     fun isGattConnected(): Boolean {
@@ -19,7 +22,13 @@ abstract class AbstractBluetoothProtocol : DefaultBluetoothGattCallback() {
         this.gatt = gatt
     }
 
-    abstract val compatibility: Array<CompatiblePeripheral>
+    fun createScanFilters(): Collection<ScanFilter> {
+        return compatiblePeripherals.map { peripheral ->
+            ScanFilter.Builder()
+                .setManufacturerData(peripheral.manufacturerId, peripheral.manufacturerData)
+                .build()
+        }
+    }
 
     abstract fun handlePacket(data: ByteArray)
